+++ src/examples/os/step_over_data.cc
@@ -42,7 +42,7 @@ sighandler_t old_handler;
  */
 void segv_handler(int sig) {
 	// printf("in segv handler\n");
-	printf("in segv handler: ptr=%p, offset is %d\n", ptr, int(ptr)%page_size);
+	printf("in segv handler: ptr=%p, offset is %ld\n", ptr, long(ptr)%page_size);
 	// call the old handler...
 	old_handler(sig);
 }
